move changelog gen to prepare-release.go
